+++ include/ignition/gazebo/components/CastShadows.hh
@@ -14,15 +14,12 @@
  * limitations under the License.
  *
  */
-#ifndef IGNITION_GAZEBO_COMPONENTS_CONTACTDATASENSOR_HH_
-#define IGNITION_GAZEBO_COMPONENTS_CONTACTDATASENSOR_HH_
+#ifndef IGNITION_GAZEBO_COMPONENTS_CASTSHADOWS_HH_
+#define IGNITION_GAZEBO_COMPONENTS_CASTSHADOWS_HH_
 
-#include <ignition/msgs/contacts.pb.h>
-#include <ignition/gazebo/components/Component.hh>
 #include <ignition/gazebo/components/Factory.hh>
-#include <ignition/gazebo/components/Serialization.hh>
+#include <ignition/gazebo/components/Component.hh>
 #include <ignition/gazebo/config.hh>
-#include <ignition/math/Vector3.hh>
 
 namespace ignition
 {
@@ -32,12 +29,11 @@ namespace gazebo
 inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
 namespace components
 {
-  /// \brief A component type that contains a list of contacts.
-  using ContactSensorData =
-      Component<msgs::Contacts,
-      class ContactSensorDataTag, serializers::MsgSerializer>;
-  IGN_GAZEBO_REGISTER_COMPONENT("ign_gazebo_components.ContactSensorData",
-                                ContactSensorData)
+  /// \brief A component used to indicate that an entity casts shadows
+  /// e.g. visual entities
+  using CastShadows = Component<bool, class CastShadowsTag>;
+  IGN_GAZEBO_REGISTER_COMPONENT("ign_gazebo_components.CastShadows",
+      CastShadows)
 }
 }
 }