+++ src/gui/plugins/scene3d/Scene3D.cc
@@ -985,6 +985,16 @@ void TextureNode::PrepareNode()
 RenderWindowItem::RenderWindowItem(QQuickItem *_parent)
   : QQuickItem(_parent), dataPtr(new RenderWindowItemPrivate)
 {
+  // FIXME(anyone) Ogre 1/2 singletons crash when there's an attempt to load
+  // this plugin twice, so shortcut here. Ideally this would be caught at
+  // Ignition Rendering.
+  static bool done{false};
+  if (done)
+  {
+    return;
+  }
+  done = true;
+
   this->setAcceptedMouseButtons(Qt::AllButtons);
   this->setFlag(ItemHasContents);
   this->dataPtr->renderThread = new RenderThread();
@@ -1116,6 +1126,18 @@ Scene3D::~Scene3D() = default;
 /////////////////////////////////////////////////
 void Scene3D::LoadConfig(const tinyxml2::XMLElement *_pluginElem)
 {
+  // FIXME(anyone) Ogre 1/2 singletons crash when there's an attempt to load
+  // this plugin twice, so shortcut here. Ideally this would be caught at
+  // Ignition Rendering.
+  static bool done{false};
+  if (done)
+  {
+    ignerr << "Only one Scene3D is supported per process at the moment."
+           << std::endl;
+    return;
+  }
+  done = true;
+
   auto renderWindow = this->PluginItem()->findChild<RenderWindowItem *>();
   if (!renderWindow)
   {
@@ -1260,11 +1282,13 @@ void Scene3D::LoadConfig(const tinyxml2::XMLElement *_pluginElem)
          << this->dataPtr->followService << "]" << std::endl;
 }
 
-
 //////////////////////////////////////////////////
 void Scene3D::Update(const UpdateInfo &_info,
     EntityComponentManager &_ecm)
 {
+  if (nullptr == this->dataPtr->renderUtil)
+    return;
+
   IGN_PROFILE("Scene3D::Update");
   if (this->dataPtr->worldName.empty())
   {
@@ -1331,6 +1355,31 @@ bool Scene3D::OnFollow(const msgs::StringMsg &_msg,
 
   _res.set_data(true);
   return true;
+}
+
+/////////////////////////////////////////////////
+void Scene3D::OnDropped(const QString &_drop)
+{
+  if (_drop.toStdString().empty())
+  {
+    ignwarn << "Dropped empty entity URI." << std::endl;
+    return;
+  }
+
+  std::function<void(const ignition::msgs::Boolean &, const bool)> cb =
+      [](const ignition::msgs::Boolean &_res, const bool _result)
+  {
+    if (!_result || !_res.data())
+      ignerr << "Error creating dropped entity." << std::endl;
+  };
+
+  // TODO(anyone) set pose according to mouse position
+  msgs::EntityFactory req;
+  req.set_sdf_filename(_drop.toStdString());
+  req.set_allow_renaming(true);
+
+  this->dataPtr->node.Request("/world/" + this->dataPtr->worldName + "/create",
+      req, cb);
 }
 
 /////////////////////////////////////////////////